feat: fix embeding links
